+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryInitProviderTest.kt
@@ -160,6 +160,7 @@ class SentryInitProviderTest {
       false,
       false,
       false,
+      false,
     )
 
     AndroidOptionsInitializer.initializeIntegrationsAndProcessors(