+++ sentry-android-distribution/src/main/java/io/sentry/android/distribution/Distribution.kt
@@ -1,3 +1,0 @@
-package io.sentry.android.distribution
-
-public class Distribution {}