+++ sentry/src/main/java/io/sentry/ExternalOptions.java
@@ -437,12 +437,10 @@ public final class ExternalOptions {
     this.sendDefaultPii = sendDefaultPii;
   }
 
-  @ApiStatus.Experimental
   public void setIgnoredCheckIns(final @Nullable List<String> ignoredCheckIns) {
     this.ignoredCheckIns = ignoredCheckIns;
   }
 
-  @ApiStatus.Experimental
   public @Nullable List<String> getIgnoredCheckIns() {
     return ignoredCheckIns;
   }
@@ -482,12 +480,10 @@ public final class ExternalOptions {
     return forceInit;
   }
 
-  @ApiStatus.Experimental
   public @Nullable SentryOptions.Cron getCron() {
     return cron;
   }
 
-  @ApiStatus.Experimental
   public void setCron(final @Nullable SentryOptions.Cron cron) {
     this.cron = cron;
   }
@@ -522,12 +518,10 @@ public final class ExternalOptions {
     return captureOpenTelemetryEvents;
   }
 
-  @ApiStatus.Experimental
   public void setEnableLogs(final @Nullable Boolean enableLogs) {
     this.enableLogs = enableLogs;
   }
 
-  @ApiStatus.Experimental
   public @Nullable Boolean isEnableLogs() {
     return enableLogs;
   }