+++ sentry/src/main/java/io/sentry/HubScopesWrapper.java
@@ -353,7 +353,6 @@ public final class HubScopesWrapper implements IHub {
     return scopes.getBaggage();
   }
 
-  @ApiStatus.Experimental
   @Override
   public @NotNull SentryId captureCheckIn(@NotNull CheckIn checkIn) {
     return scopes.captureCheckIn(checkIn);
@@ -370,7 +369,6 @@ public final class HubScopesWrapper implements IHub {
     return scopes.captureReplay(replay, hint);
   }
 
-  @ApiStatus.Experimental
   @Override
   public @NotNull ILoggerApi logger() {
     return scopes.logger();