+++ sentry/src/main/java/io/sentry/MonitorContexts.java
@@ -6,11 +6,9 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-@ApiStatus.Experimental
 public final class MonitorContexts extends ConcurrentHashMap<String, Object>
     implements JsonSerializable {
   private static final long serialVersionUID = 3987329379811822556L;